refractoered the util functions out
