+++ render_first_frame.py
@@ -4,11 +4,9 @@
 
 from __future__ import annotations
 
-import json
-import math
 from argparse import ArgumentParser
 from pathlib import Path
-from typing import Sequence, TextIO
+from typing import TextIO
 
 import traceback
 from datetime import datetime
@@ -20,9 +18,8 @@ import torch
 from arguments import PipelineParams
 from gaussian_renderer import render_or
 from scene import GaussianModel
-from scene.cameras import MiniCam
 
-from utils.graphics_utils import getProjectionMatrix
+from utils.render_utils import build_perspective_camera, load_raster_world_points_only
 
 
 JSON_DIR = Path("/home/tianhang/habitatbuild/habitat-sim/navdata_multi_terrain_pixelspace_with_gs")
@@ -30,87 +27,6 @@ PLY_PATH = Path("/home/tianhang/NavDP/navdp_api/gaussian_splatting/data/floor30g
 DEFAULT_ERROR_LOG = JSON_DIR.parent / "errors.log"
 EPS = 1e-6
 CAMERA_HEIGHT_OFFSET = 1.3
-
-
-def load_raster_world_points(
-    json_path: Path,
-) -> list[dict]:
-    """Extract raster_world points."""
-
-    with json_path.open("r", encoding="utf-8") as fh:
-        payload = json.load(fh)
-
-    path_payload = payload.get("path", {})
-    raster_world = path_payload.get("raster_world")
-    if not raster_world:
-        raise ValueError(f"Missing raster_world in {json_path}")
-
-    return raster_world
-
-
-def build_look_at(eye: np.ndarray, target: np.ndarray, up: np.ndarray) -> np.ndarray:
-    """Construct a right-handed look-at view matrix."""
-
-    forward = target - eye
-    forward_norm = np.linalg.norm(forward)
-    if forward_norm < EPS:
-        raise ValueError("Camera target too close to position; cannot build view matrix.")
-    forward /= forward_norm
-
-    right = np.cross(up, forward)
-    right_norm = np.linalg.norm(right)
-    if right_norm < EPS:
-        fallback = np.array([1.0, 0.0, 0.0], dtype=np.float32)
-        if abs(np.dot(fallback, forward)) > 0.99:
-            fallback = np.array([0.0, 1.0, 0.0], dtype=np.float32)
-        right = np.cross(fallback, forward)
-        right_norm = np.linalg.norm(right)
-    right /= max(right_norm, EPS)
-
-    true_up = np.cross(forward, right)
-
-    view = np.eye(4, dtype=np.float32)
-    view[0, :3] = right
-    view[1, :3] = true_up
-    view[2, :3] = forward
-    view[:3, 3] = -view[:3, :3] @ eye
-    return view
-
-
-def build_perspective_camera(
-    position: np.ndarray,
-    target: np.ndarray,
-    width: int,
-    height: int,
-    fov_deg: float,
-    znear: float,
-    zfar: float,
-    device: torch.device,
-) -> MiniCam:
-    """Generate a MiniCam with perspective projection."""
-
-    if height <= 0 or width <= 0:
-        raise ValueError("Camera resolution must be positive.")
-
-    fovy = math.radians(fov_deg)
-    aspect = width / max(height, 1)
-    fovx = 2.0 * math.atan(math.tan(fovy * 0.5) * aspect)
-
-    view = build_look_at(position, target, np.array([0.0, 0.0, 1.0], dtype=np.float32))
-    world_view = torch.from_numpy(view).to(device).transpose(0, 1)
-    projection = getProjectionMatrix(znear=znear, zfar=zfar, fovX=fovx, fovY=fovy).to(device).transpose(0, 1)
-    full_proj = (world_view.unsqueeze(0) @ projection.unsqueeze(0)).squeeze(0)
-
-    return MiniCam(
-        width=width,
-        height=height,
-        fovy=fovy,
-        fovx=fovx,
-        znear=znear,
-        zfar=zfar,
-        world_view_transform=world_view,
-        full_proj_transform=full_proj,
-    )
 
 
 def render_path_first_frame(
@@ -130,16 +46,15 @@ def render_path_first_frame(
 ) -> None:
     """Render first frame for a raster_world trajectory."""
 
-    raster_world = load_raster_world_points(json_path)
+    raster_world = load_raster_world_points_only(json_path)
     if len(raster_world) < 2:
         raise ValueError(f"Need at least two points in {json_path}")
 
     first = raster_world[0]
     second = raster_world[1]
 
-    position = np.array([first["x"], first["y"], first["z"] + CAMERA_HEIGHT_OFFSET], dtype=np.float32)
-
-    target = np.array([second["x"], second["y"], second["z"] + CAMERA_HEIGHT_OFFSET], dtype=np.float32)
+    position = np.array([first[0], first[1], first[2] + CAMERA_HEIGHT_OFFSET], dtype=np.float32)
+    target = np.array([second[0], second[1], second[2] + CAMERA_HEIGHT_OFFSET], dtype=np.float32)
 
     top_down_map_dir = JSON_DIR / "gs_obs"
     top_down_map_dir.mkdir(parents=True, exist_ok=True)
